Implemented MultiTenancy scaffold in Appointment Service

+++ appointment-service/src/main/java/com/ams/appointment_service/service/StaffService.java
@@ -6,7 +6,7 @@ import com.ams.appointment_service.mapper.AppointmentMapper;
 import com.ams.appointment_service.mapper.UpdateResponseMapper;
 import com.ams.appointment_service.model.entities.Appointment;
 import com.ams.appointment_service.model.entities.StaffScheduleSnapshot;
-import com.ams.appointment_service.model.constant.StaffConfirmStatus;
+import com.ams.appointment_service.model.constant.AppointmentStatus;
 import com.ams.appointment_service.repository.AppointmentRepository;
 import com.ams.appointment_service.repository.StaffScheduleSnapshotRepository;
 import lombok.AllArgsConstructor;
@@ -20,6 +20,7 @@ public class StaffService {
 
     private final StaffScheduleSnapshotRepository staffRepository;
     private final AppointmentRepository appointmentRepository;
+    private final NotificationService notificationService;
 
     public Optional<StaffScheduleSnapshot> getStaffSnapShot(UUID staffId) {
         return staffRepository.findById(staffId);
@@ -29,12 +30,22 @@ public class StaffService {
         staffRepository.save(snapshot);
     }
 
-    public void confirmAppointment(long appointmentId) {
+    public void confirmAppointment(long appointmentId, boolean accepted) {
         Optional<Appointment> optional = appointmentRepository.findById(appointmentId);
         if (optional.isEmpty()) return;
         Appointment appointment = optional.get();
-        appointment.setStatus(StaffConfirmStatus.APPROVED);
+        appointment.setStaffStatus(accepted ? AppointmentStatus.APPROVED : AppointmentStatus.DECLINED);
         appointmentRepository.save(appointment);
+        notificationService.notifyCustomerAppointmentBookingWasConfirmed(appointment);
+    }
+
+    public void confirmAppointmentReschedule(long appointmentId, boolean accepted) {
+        Optional<Appointment> optional = appointmentRepository.findById(appointmentId);
+        if (optional.isEmpty()) return;
+        Appointment appointment = optional.get();
+        appointment.setStaffStatus(accepted ? AppointmentStatus.APPROVED : AppointmentStatus.DECLINED);
+        appointmentRepository.save(appointment);
+        notificationService.notifyCustomerReschedulingWasComFirm(appointment);
     }
 
     public UpdateDTO editAppointment(UpdateDTO updateRequest) {
@@ -45,13 +56,20 @@ public class StaffService {
         appointment.setDate(updateRequest.getDate());
         appointment.setStartTime(updateRequest.getStartTime());
         appointment.setEndTime(updateRequest.getEndTime());
-        appointment.setStatus(updateRequest.getStatus());
+        appointment.setStaffStatus(updateRequest.getStatus());
         var updatedAppointment = appointmentRepository.save(appointment);
+        notificationService.notifyCustomerAppointmentWasRescheduled(appointment);
         return UpdateResponseMapper.toDTO(updatedAppointment);
     }
 
     public void deleteAppointment(long appointmentId, UUID staffId) {
-        appointmentRepository.deleteByIdAndStaffId(appointmentId, staffId);
+        Optional<Appointment> optional = appointmentRepository.findById(appointmentId);
+        if (optional.isEmpty()) return;
+        Appointment appointment = optional.get();
+        long aptId = appointmentRepository.deleteByIdAndStaffId(appointmentId, staffId);
+        if (aptId != 0) {
+            notificationService.notifyCustomerAppointmentHasBeenCancelled(appointment);
+        }
     }
 
     public List<AppointmentResponseDTO> getAllAppointmentForAStaff(UUID staffId) {
@@ -59,8 +77,6 @@ public class StaffService {
         return staff.map(staffScheduleSnapshot -> staffScheduleSnapshot.getAppointments()
                 .stream().map(AppointmentMapper::toDTO).toList()).orElse(Collections.emptyList());
     }
-
-
 
 }
